Generate a random username on first logon
The username is a combination of cowboy-themed adjectives and nouns,
    plus a number from 0 to 100.
